rename Card to CardType, add Card
