+++ client/src/components/CardPile/CardPile.tsx
@@ -1,15 +1,12 @@
 import React, { FC } from 'react'
 import { CSSTransition, TransitionGroup } from 'react-transition-group'
 
-import Card from '../Card'
-import { Card as CardType } from '../../types/Card'
+import CardComponent from '../Card'
+import { Card } from '../../types/Card'
 import './CardPile.scss'
 
 export interface CardPileProps {
-  cards: {
-    type: CardType
-    id: string
-  }[]
+  cards: Card[]
   className?: string
 }
 
@@ -27,7 +24,7 @@ const CardPile: FC<CardPileProps> = ({ cards, className }) => {
           timeout={{ appear: 250, enter: 250, exit: 0 }}
           classNames='card-animation'
         >
-          <Card type={type} className='CardPile__Card' />
+          <CardComponent type={type} className='CardPile__Card' />
         </CSSTransition>
       ))}
     </TransitionGroup>
