+++ client/src/types/CardType.ts
@@ -1,0 +1,5 @@
+export enum CardType {
+  EGG,
+  CHICKEN,
+  FOX
+}
